refactor: split macro handling
